+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -68,7 +68,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc3ce289afsc6egz69qexm0b"
+        "conv_01kc3pkd45a38qw2m76ej9qjt8"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """