+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -98,7 +98,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc3ce28gp57afvf8gwe8s8b0"
+        "conv_01kc3pkd4709ft9xrmq3j52reh"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """