+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1106,7 +1106,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-12-10T05:41:06.457Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-12-10T08:38:47.231Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1158,7 +1158,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-12-10T05:41:06.456Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-12-10T08:38:47.231Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1210,7 +1210,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-12-10T05:41:06.456Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-12-10T08:38:47.230Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1289,5 +1289,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-10T05:40:54.670Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-10T08:38:35.196Z")
     r"""The date and time the resource was last updated"""