+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KC3CE2NSA6BQZJN9RDRN7AMH"
+        "01KC3PKDJTEYC0C55T85VZ0879"
     )
     r"""The unique identifier of the data source"""