+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -150,7 +150,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc3ce28pyjrqge73b610b30g"
+        "conv_01kc3pkd49angwcqy3ngdn6tsf"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """