+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveconversationop.py
@@ -110,7 +110,7 @@ class RetrieveConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc3ce28ryty8dg5bnwyth1za"
+        "conv_01kc3pkd4bbja31h18qhpca8e3"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """