+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -80,5 +80,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-10T05:40:54.670Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-10T08:38:35.196Z")
     r"""The date and time the resource was last updated"""