+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KC3CE2NRQNYAH3JN6XW199SS"
+        "01KC3PKDJSZST0Y8QCEQQ696QN"
     )
     r"""The unique identifier of the data source"""