+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -146,7 +146,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kc3ce28z7j1rxx1axnrez39h"
+        "conv_01kc3pkd4chwb8emxha7g9cr7k"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """