+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KC3CE2NV5ACX6KTJHFYKPC7Z"
+        "01KC3PKDJWMN3CS0Q6GGVTZG8C"
     )
     r"""The unique identifier of the data source"""